+++ NEWS.md
@@ -4,6 +4,7 @@
 * New utility function `cut_quantile` for cutting a numeric vector into quantiles.
 * New utility function `cat_with_newline` for concatenating and printing with newline.
 * New check function `check_proportion` which checks for a single proportion.
+* Better legends on the genes barplot and the correlation heatmap.
 
 # hermes 0.1.1